+++ docs/reference/change-log.md
@@ -2,9 +2,14 @@
 
 
 ## Unofficial Release
-###### Includes changes up to 2026-02-11
+###### Includes changes up to 2026-02-17
+
+#### Not Specified
+* Add testing packages to phlynx by [@jmdowrick](https://github.com/jmdowrick) in [#308](https://github.com/physiomelinks/phlynx/pull/308)
+
 
 #### Fixed Bugs
+* Fix bugs related to system dialog by [@jmdowrick](https://github.com/jmdowrick) in [#307](https://github.com/physiomelinks/phlynx/pull/307)
 * Fix order of columns in parameter export by [@jmdowrick](https://github.com/jmdowrick) in [#294](https://github.com/physiomelinks/phlynx/pull/294)
 * Fix circulatory autogen export naming by [@jmdowrick](https://github.com/jmdowrick) in [#291](https://github.com/physiomelinks/phlynx/pull/291)
 * Fix parameter export format by [@jmdowrick](https://github.com/jmdowrick) in [#287](https://github.com/physiomelinks/phlynx/pull/287)